v1/accounts/managers/user_manager.py
from django.contrib.auth.models import BaseUserManager


class UserManager(BaseUserManager):
    use_in_migrations = True

    def _create_user(self, email, password, **kwargs):
        email = self.normalize_email(email)
        type_choice = kwargs.pop('type_choice', "4")
        user = self.model(
            email=email,
            is_active=True,
            type_choice=type_choice,
            **kwargs
        )
        user.set_password(password)
        user.save(using=self._db)
        return user

    def create_user(self, email, password=None, **extra_fields):
        return self._create_user(email, password, **extra_fields)

    def create_superuser(self, email, password, **extra_fields):
        return self._create_user(email, password, is_active=True, type_choice=1, **extra_fields)
